// file-manager/src/app/core/models/client.model.ts
export interface Client {
    id?: number;
    name: string;
    mobileNo?: string;
    email?: string;
    address?: string;
    villageId?: number;
    villageName?: string;
    talukaName?: string;
    districtName?: string;
    stateName?: string;
    countryName?: string;
    createdAt?: string;
}